allow build on non-unix systems
This is required to use this library on windows. Even if only the verification part of the library is required, the build fails if ioctl / raw syscalls are used in tdx/tdx.go
